* object.c (array_slice2): use a cast to "size_t".

+++ src/object.c
@@ -792,7 +792,7 @@ array_slice2(struct narray *array, int start, int length)
   }
 
   if (length > 0) {
-    memmove(data, data + start * base, length * base);
+    memmove(data, data + start * base, (size_t) length * base);
   }
 
   array->num = length;
